requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.8
langchain-openai>=0.0.5
langgraph>=0.0.20
google-generativeai>=0.3.0
openai>=1.3.0
pandas>=2.0.0
duckdb>=0.8.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
plotly>=5.17.0
reportlab>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.23.0
openpyxl>=3.1.0
